tests/wpt/css-tests/selectors-3_dev/xhtml1print/css3-selectors-lang-031.xht

<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
<html lang="en" xmlns="http://www.w3.org/1999/xhtml"><head>
<meta charset="utf-8" />
<title>[lang|="az-Arab-IR"], lang="az-Arab-IR"</title>
<link href="mailto:ishida@w3.org" rel="author" title="Richard Ishida" />
<link href="http://www.w3.org/TR/css3-selectors/#attribute-selectors" rel="help" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style type="text/css">
.test div { width: 50px; }
 		#colonlangcontroltest { color: red; font-weight: bold; width: 400px; }
		#colonlangcontroltest[lang|=xx] { display:none; }
#box[lang|='az-Arab-IR'] { width: 100px; }
#relevance[lang|='yy'] { display:none; }
</style>
</head>
<body>



<div class="test"><div lang="az-Arab-IR" id="box">&#xA0;</div></div>
<p lang="xx" id="colonlangcontroltest">This test failed because it relies on [lang|=..] for results, but [lang|=..] is not supported by this browser.


<!--Notes:
This tests a detail related to [lang|=..] support. If [lang|=..] is not supported, a message will appear and the test will fail.
-->
<script>
test(function() {
assert_equals(document.getElementById('colonlangcontroltest').offsetWidth, 0)
assert_equals(document.getElementById('box').offsetWidth, 100);
}, "A lang|= value with language, script and region subtags will match a lang attribute value with the same language, script and region subtags.");
</script>

</p><div id="log"></div>



</body></html>
